Minor weapon refactor

+++ Assets/Gun Pistol Package/Associated Gun scripts/Interfaces/IDamagable.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamagable 
 {
-    public abstract void TakeDamage(int Amount);
+    public abstract void TakeDamage(float Amount);
     
 }
